Move common cue files generation into plugins
